Move pom file format method to file operations
